requirements.txt
redis~=3.2.100
rq~=1.7.0
flask~=1.1.2
pillow~=8.1.0
python-dotenv~=0.15.0
numpy~=1.19.5
scipy~=1.6.0
tqdm~=4.56.0
torchvision~=0.8.2
requests~=2.25.1
fastai~=1.0.51
ffmpeg~=4.3.1
matplotlib~=3.3.3
tensorboardx~=2.1
setuptools~=51.3.3
PIL